pyproject.toml
[project]
name = "furia-gemini-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "google-genai>=1.13.0",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.1.0",
    "ruff>=0.11.8",
]
